requirement.txt
streamlit
torch
torchvision
opencv-python-headless
numpy
pandas
Pillow
albumentations
timm